feat: 다량의 소켓 연결을 위한 실시간 차량 현황 페이지 분리 (#10)
