+++ app/(main)/live/page.tsx
@@ -1,74 +1,25 @@
 'use client'
 
-import { Tooltip } from '@mantine/core'
-import { useSearchParams } from 'next/navigation'
-import { useRef, useState } from 'react'
+import { useEffect, useRef } from 'react'
 
 import MapSection from '@/app/(main)/live/components/MapSection'
-import RouteSearchSection from '@/app/(main)/live/components/RouteSearchSection'
-import LiveMode from '@/components/common/LiveMode'
 import { useLiveRoute } from '@/hooks/useLiveRoute'
-import { useQueryParams } from '@/hooks/useQueryParams'
-import { vars } from '@/styles/theme.css'
-import { LatLng } from '@/types/map'
 
 import * as styles from './styles.css'
 
 const LivePage = () => {
-    const [routes, setRoutes] = useState<LatLng[]>([])
-    const [isMapLoaded, setIsMapLoaded] = useState(false)
-
-    const { currentLocation, startLiveTracking, stopLiveTracking } = useLiveRoute()
+    const { currentLocations, connectSocket, disconnectSocket } = useLiveRoute()
 
     const mapRef = useRef<kakao.maps.Map>(null)
 
-    const searchParams = useSearchParams()
-
-    const { addQuery, removeQuery } = useQueryParams()
-
-    const live = searchParams.get('live') === 'true'
-
-    const handleTrackingToggle = (isClicked: boolean) => {
-        if (isClicked) {
-            addQuery('tracking', 'true')
-        } else {
-            removeQuery('tracking')
-        }
-    }
-
-    // console.log('currentLocation', currentLocation)
+    useEffect(() => {
+        connectSocket()
+        return () => disconnectSocket()
+    }, [])
 
     return (
         <div className={styles.container}>
-            {live && (
-                <Tooltip
-                    label={'차량 위치 자동 추적'}
-                    color={vars.colors.gray[800]}
-                    arrowSize={6}
-                    withArrow
-                    arrowPosition='side'
-                >
-                    <div className={styles.trackingButton}>
-                        <LiveMode disabled={!live} onChange={handleTrackingToggle} />
-                    </div>
-                </Tooltip>
-            )}
-            <RouteSearchSection
-                mapRef={mapRef}
-                onRoutesChange={setRoutes}
-                startLiveTracking={startLiveTracking}
-                stopLiveTracking={stopLiveTracking}
-            />
-
-            <MapSection
-                mapRef={mapRef}
-                routes={routes}
-                // initialLiveRoute={initialLiveRoute!}
-                currentLocation={currentLocation!}
-                isMapLoaded={isMapLoaded}
-                onRoutesChange={setRoutes}
-                onLoad={() => setIsMapLoaded(true)}
-            />
+            <MapSection mapRef={mapRef} currentLocations={currentLocations!} />
         </div>
     )
 }
